Refactored simulation_class to simulation and made adjustments to rpaths
