+++ CMakeLists.txt
@@ -54,9 +54,9 @@ ELSE()
         SET(INSTALL_INCLUDEDIR ${SKBUILD_HEADERS_DIR}) 
         SET(INSTALL_PYPROJ ${SKBUILD_PLATLIB_DIR}/${SKBUILD_PROJECT_NAME}) 
         IF (APPLE)
-            SET(CMAKE_INSTALL_RPATH "@loader_path;@loader_path/../lib")
+            SET(CMAKE_INSTALL_RPATH "@loader_path;@loader_path/../lib;${CMAKE_BINARY_DIR}/bin")
         ELSEIF (LINUX)
-            SET(CMAKE_INSTALL_RPATH "@ORIGIN;@ORIGIN/../lib")
+            SET(CMAKE_INSTALL_RPATH "@ORIGIN;${INSTALL_LIBDIR};${CMAKE_BINARY_DIR}/bin")
         ENDIF ()
     ELSE ()
         SET(INSTALL_PYPROJ ${PY})
@@ -65,6 +65,11 @@ ELSE()
         SET(INSTALL_INCLUDEDIR ${CMAKE_INSTALL_INCLUDEDIR})
     ENDIF ()
     
+    MESSAGE(STATUS "INSTALL_BINDIR: ${INSTALL_BINDIR}")
+    MESSAGE(STATUS "INSTALL_LIBDIR: ${INSTALL_LIBDIR}")
+    MESSAGE(STATUS "INSTALL_INCLUDEDIR: ${INSTALL_INCLUDEDIR}")
+    MESSAGE(STATUS "INSTALL_PYPROJ: ${INSTALL_PYPROJ}")
+    MESSAGE(STATUS "CMAKE_INSTALL_RPATH: ${CMAKE_INSTALL_RPATH}")
     SET(CMAKE_INSTALL_RPATH_USE_LINK_PATH TRUE)
 
     # Have the .mod files placed in the include folder
