+++ swiftest/__init__.py
@@ -1,5 +1,5 @@
 """
-Copyright 2022 - David Minton, Carlisle Wishard, Jennifer Pouplin, Jake Elliott, & Dana Singh
+Copyright 2024 - The Minton Group at Purdue University
 This file is part of Swiftest.
 Swiftest is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License 
 as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
@@ -10,5 +10,5 @@ If not, see: https://www.gnu.org/licenses.
 """
 
 from .constants import *
-from .simulation_class import Simulation
+from .simulation import Simulation
 from .shgrav import clm_from_ellipsoid, clm_from_relief
